+++ packages/zenrouter_file_generator/example/lib/main_deferred.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:zenrouter_file_generator_example/flutter_scan.dart';
 import 'routes/routes.zen.dart'
     deferred as routes
-    hide AppCoordinatorNav, AppCoordinatorGetter;
+    hide AppCoordinatorNav, AppCoordinatorGetter, AppCoordinatorNavContext;
 
 void main() {
   runApp(FlutterScan(child: const MyApp()));